Disentangle core expression class from dataframe specific code
