+++ dask/dataframe/dask_expr/_expr.py
@@ -33,9 +33,9 @@ from dask.dataframe.core import (
 )
 from dask.dataframe.dask_expr import _core as core
 from dask.dataframe.dask_expr._util import (
+    _BackendData,
     _calc_maybe_new_divisions,
     _convert_to_list,
-    _tokenize_deterministic,
     _tokenize_partial,
     is_scalar,
 )
@@ -48,7 +48,7 @@ from dask.dataframe.utils import (
     raise_on_meta_error,
     valid_divisions,
 )
-from dask.tokenize import normalize_token
+from dask.tokenize import _tokenize_deterministic, normalize_token
 from dask.typing import Key, no_default
 from dask.utils import (
     M,
@@ -105,6 +105,74 @@ class Expr(core.Expr):
     @property
     def nbytes(self):
         return NBytes(self)
+
+    def _tree_repr_lines(self, indent=0, recursive=True):
+        header = funcname(type(self)) + ":"
+        lines = []
+        for i, op in enumerate(self.operands):
+            if isinstance(op, Expr):
+                if recursive:
+                    lines.extend(op._tree_repr_lines(2))
+            else:
+                if isinstance(op, _BackendData):
+                    op = op._data
+
+                # TODO: this stuff is pandas-specific
+                if isinstance(op, pd.core.base.PandasObject):
+                    op = "<pandas>"
+                elif is_dataframe_like(op):
+                    op = "<dataframe>"
+                elif is_index_like(op):
+                    op = "<index>"
+                elif is_series_like(op):
+                    op = "<series>"
+                elif is_arraylike(op):
+                    op = "<array>"
+                header = self._tree_repr_argument_construction(i, op, header)
+
+        lines = [header] + lines
+        lines = [" " * indent + line for line in lines]
+
+        return lines
+
+    def _operands_for_repr(self):
+        to_include = []
+        for param, operand in zip(self._parameters, self.operands):
+            if isinstance(operand, Expr) or (
+                not isinstance(operand, (pd.Series, pd.DataFrame))
+                and operand != self._defaults.get(param)
+            ):
+                to_include.append(f"{param}={operand!r}")
+        return to_include
+
+    def __getattr__(self, key):
+        try:
+            return object.__getattribute__(self, key)
+        except AttributeError as err:
+            if key.startswith("_meta"):
+                # Avoid a recursive loop if/when `self._meta*`
+                # produces an `AttributeError`
+                raise RuntimeError(
+                    f"Failed to generate metadata for {self}. "
+                    "This operation may not be supported by the current backend."
+                )
+
+            # Allow operands to be accessed as attributes
+            # as long as the keys are not already reserved
+            # by existing methods/properties
+            _parameters = type(self)._parameters
+            if key in _parameters:
+                idx = _parameters.index(key)
+                return self.operands[idx]
+            if is_dataframe_like(self._meta) and key in self._meta.columns:
+                return self[key]
+
+            link = "https://github.com/dask-contrib/dask-expr/blob/main/README.md#api-coverage"
+            raise AttributeError(
+                f"{err}\n\n"
+                "This often means that you are attempting to use an unsupported "
+                f"API function. Current API coverage is documented here: {link}."
+            )
 
     def __getitem__(self, other):
         if isinstance(other, Expr):
@@ -590,7 +658,7 @@ class MapPartitions(Blockwise):
         "token",
         "kwargs",
     ]
-    _defaults = {
+    _defaults: dict = {
         "kwargs": None,
         "align_dataframes": True,
         "parent_meta": None,
@@ -842,7 +910,7 @@ class MapOverlap(MapPartitions):
         "token",
         "kwargs",
     ]
-    _defaults = {
+    _defaults: dict = {
         "meta": None,
         "enfore_metadata": True,
         "transform_divisions": True,
@@ -1420,7 +1488,7 @@ class ToTimestamp(Elemwise):
 
 class CombineSeries(Elemwise):
     _parameters = ["frame", "other", "func", "fill_value"]
-    _defaults = {"fill_value": None}
+    _defaults: dict = {"fill_value": None}
     operation = M.combine
 
     @functools.cached_property
@@ -1643,14 +1711,14 @@ class ToFrame(Elemwise):
 
 class ToFrameIndex(ToFrame):
     _parameters = ["frame", "index", "name"]
-    _defaults = {"name": no_default, "index": True}
+    _defaults = {"name": no_default, "index": True}  # type: ignore
     _keyword_only = ["name", "index"]
     operation = M.to_frame
     _filter_passthrough = True
 
 
 class ToSeriesIndex(ToFrameIndex):
-    _defaults = {"name": no_default, "index": None}
+    _defaults = {"name": no_default, "index": None}  # type: ignore
     operation = M.to_series
     _preserves_partitioning_information = True
 
@@ -1703,7 +1771,7 @@ class Apply(Elemwise):
     """A good example of writing a less-trivial blockwise operation"""
 
     _parameters = ["frame", "function", "args", "meta", "kwargs"]
-    _defaults = {"args": (), "kwargs": {}}
+    _defaults = {"args": (), "kwargs": {}}  # type: ignore
     operation = M.apply
 
     @functools.cached_property
@@ -1935,7 +2003,7 @@ class Assign(Elemwise):
 
 class Eval(Elemwise):
     _parameters = ["frame", "_expr", "expr_kwargs"]
-    _defaults = {"expr_kwargs": {}}
+    _defaults = {"expr_kwargs": {}}  # type: ignore
     _keyword_only = ["expr_kwargs"]
     operation = M.eval
 
@@ -3046,7 +3114,7 @@ def optimize_until(expr: Expr, stage: core.OptimizerStage) -> Expr:
         return result
 
     # Simplify
-    expr = result.simplify()  # type: ignore
+    expr = result.simplify()
     if stage == "simplified-logical":
         return expr
 
@@ -3056,12 +3124,12 @@ def optimize_until(expr: Expr, stage: core.OptimizerStage) -> Expr:
         return expr
 
     # Lower
-    expr = expr.lower_completely()  # type: ignore
+    expr = expr.lower_completely()
     if stage == "physical":
         return expr
 
     # Simplify again
-    expr = expr.simplify()  # type: ignore
+    expr = expr.simplify()
     if stage == "simplified-physical":
         return expr
