+++ dask/dataframe/dask_expr/_util.py
@@ -12,7 +12,7 @@ from packaging.version import Version
 from dask import config, is_dask_collection
 from dask.dataframe._compat import is_string_dtype
 from dask.dataframe.core import is_dataframe_like, is_series_like
-from dask.tokenize import normalize_token, tokenize
+from dask.tokenize import _tokenize_deterministic, normalize_token
 from dask.utils import get_default_shuffle_method
 
 K = TypeVar("K", bound=Hashable)
@@ -95,11 +95,6 @@ def is_scalar(x):
     return not isinstance(x, Expr)
 
 
-def _tokenize_deterministic(*args, **kwargs) -> str:
-    # Utility to be strict about deterministic tokens
-    return tokenize(*args, ensure_deterministic=True, **kwargs)
-
-
 def _tokenize_partial(expr, ignore: list | None = None) -> str:
     # Helper function to "tokenize" the operands
     # that are not in the `ignore` list
@@ -145,7 +140,7 @@ class _BackendData:
 
     @functools.cached_property
     def _token(self):
-        from dask.dataframe.dask_expr._util import _tokenize_deterministic
+        from dask.tokenize import _tokenize_deterministic
 
         return _tokenize_deterministic(self._data)
