+++ dask/dataframe/dask_expr/io/_delayed.py
@@ -12,11 +12,11 @@ from dask.dataframe.dask_expr._expr import (
     DelayedsExpr,
     PartitionsFiltered,
 )
-from dask.dataframe.dask_expr._util import _tokenize_deterministic
 from dask.dataframe.dask_expr.io import BlockwiseIO
 from dask.dataframe.dispatch import make_meta
 from dask.dataframe.utils import check_meta, pyarrow_strings_enabled
 from dask.delayed import Delayed, delayed
+from dask.tokenize import _tokenize_deterministic
 from dask.typing import Key
 
 if TYPE_CHECKING:
