+++ dask/dataframe/dask_expr/io/parquet.py
@@ -45,7 +45,7 @@ from dask.dataframe.dask_expr._expr import (
     determine_column_projection,
 )
 from dask.dataframe.dask_expr._reductions import Len
-from dask.dataframe.dask_expr._util import _convert_to_list, _tokenize_deterministic
+from dask.dataframe.dask_expr._util import _convert_to_list
 from dask.dataframe.dask_expr.io import BlockwiseIO, PartitionsFiltered
 from dask.dataframe.dask_expr.io.io import FusedParquetIO
 from dask.dataframe.io.parquet.core import (
@@ -60,7 +60,7 @@ from dask.dataframe.io.parquet.core import (
 from dask.dataframe.io.parquet.utils import _split_user_options
 from dask.dataframe.io.utils import _is_local_fs
 from dask.delayed import delayed
-from dask.tokenize import normalize_token, tokenize
+from dask.tokenize import _tokenize_deterministic, normalize_token, tokenize
 from dask.typing import Key
 from dask.utils import apply, funcname, natural_sort_key, parse_bytes, typename
 
@@ -1670,8 +1670,8 @@ class _DNF:
             ):
                 # Simple dict to make sure field comes first in filter
                 flip = {LE: GE, LT: GT, GE: LE, GT: LT}
-                op = predicate_expr  # type: ignore
-                op = flip.get(op, op)._operator_repr  # type: ignore
+                op = predicate_expr
+                op = flip.get(op, op)._operator_repr
                 column = predicate_expr.right.columns[0]
                 value = predicate_expr.left
                 _filters = (column, op, value)
